+++ src/main/java/model/Credencial.java
@@ -1,19 +1,19 @@
 package model;
 
 public class Credencial {
-    private final int idCredencial;
+    private final long idCredencial;
     private final String email;
     private final String senha;
     private final boolean ativo;
 
-    public Credencial(int idCredencial, String email, String senha, boolean ativo) {
+    public Credencial(long idCredencial, String email, String senha, boolean ativo) {
         this.idCredencial = idCredencial;
         this.email = email;
         this.senha = senha;
         this.ativo = ativo;
     }
 
-    public int getIdCredencial() {
+    public long getIdCredencial() {
         return idCredencial;
     }
